requirements.txt
a2a-sdk>=0.2.1
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
openai>=1.0.0
anthropic>=0.7.0
google-generativeai>=0.3.0
cohere>=4.0.0
requests>=2.31.0
aiohttp>=3.9.0
pydantic>=2.5.0 